requirements.txt
asgiref==3.6.0
backports.zoneinfo==0.2.1;python_version<"3.9"
boto3==1.26.138
botocore==1.29.138
cffi==1.15.1
cryptography==40.0.2
Django==4.2.1
django-environ==0.10.0
django-storages==1.13.2
gunicorn==20.1.0
jmespath==1.0.1
mysqlclient==2.1.1
Pillow==9.5.0
protobuf==3.20.3
pycparser==2.21
python-dateutil==2.8.2
s3transfer==0.6.1
six==1.16.0
sqlparse==0.4.4
urllib3==1.26.15
